+++ fabric/src/main/java/de/maxhenkel/voicechat/permission/FabricPermissionManager.java
@@ -1,5 +1,8 @@
 package de.maxhenkel.voicechat.permission;
 
+import de.maxhenkel.voicechat.Voicechat;
+import me.lucko.fabric.api.permissions.v0.Permissions;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.server.level.ServerPlayer;
 
 public class FabricPermissionManager extends PermissionManager {
@@ -9,6 +12,9 @@ public class FabricPermissionManager extends PermissionManager {
         return new Permission() {
             @Override
             public boolean hasPermission(ServerPlayer player) {
+                if (isFabricPermissionsAPILoaded()) {
+                    return Permissions.check(player, node, type.hasPermission(player));
+                }
                 return type.hasPermission(player);
             }
 
@@ -17,6 +23,18 @@ public class FabricPermissionManager extends PermissionManager {
                 return type;
             }
         };
+    }
+
+    private static Boolean loaded;
+
+    private static boolean isFabricPermissionsAPILoaded() {
+        if (loaded == null) {
+            loaded = FabricLoader.getInstance().isModLoaded("fabric-permissions-api-v0");
+            if (loaded) {
+                Voicechat.LOGGER.info("Using Fabric Permissions API");
+            }
+        }
+        return loaded;
     }
 
 }